Refresh tokens are revoked when used
Adds new configuration variable `refresh_token_revoked_in`. If you need
refresh tokens to be revoked at some point in the future (to allow room
for network failures, for instance) you can configure
`refresh_token_revoked_in` to a number of seconds into the future. It's
`0` by default.

This changeset adds `previous_refresh_token` attribute to `AccessToken`.
It is updated when a new access_token is requested, and it is revoked
when authenticating an access token with it.